Cleanup some un-used & sort types, split files
