+++ packages/bot/src/index.ts
@@ -8,8 +8,9 @@ export * from './constants.js'
 export * from './handlers/index.js'
 export * from './handlers.js'
 export * from './helpers.js'
-export * from './optionalize.js'
 export * from './transformers/index.js'
 export * from './transformers.js'
 export * from './typings.js'
+export * from './events.js'
+export * from './desiredProprieties.js'
 export * from './utils.js'
