+++ packages/bot/src/transformers/index.ts
@@ -42,6 +42,7 @@ export * from './user.js'
 export * from './voiceRegion.js'
 export * from './voiceState.js'
 export * from './webhook.js'
+export * from './subscription.js'
 export * from './welcomeScreen.js'
 export * from './widget.js'
 export * from './widgetSettings.js'
