+++ packages/bot/src/transformers/reverse/component.ts
@@ -1,30 +1,134 @@
-import type { Bot, Component } from '../../index.js'
-import type { DiscordComponent } from '../../typings.js'
+import { type DiscordButtonComponent, type DiscordMessageComponent, MessageComponentTypes, type TextStyles } from '@discordeno/types'
+import {
+  type Bot,
+  type ButtonStyles,
+  type Component,
+  type DiscordActionRow,
+  type DiscordInputTextComponent,
+  type DiscordSelectMenuComponent,
+} from '../../index.js'
 
-export function transformComponentToDiscordComponent(bot: Bot, payload: Component): DiscordComponent {
+export function transformComponentToDiscordComponent(bot: Bot, payload: Component): DiscordMessageComponent {
+  // This switch should include all cases
+  switch (payload.type) {
+    case MessageComponentTypes.ActionRow:
+      return transformActionRow(bot, payload)
+    case MessageComponentTypes.Button:
+      return transformButtonComponent(bot, payload)
+    case MessageComponentTypes.InputText:
+      return transformInputTextComponent(bot, payload)
+    case MessageComponentTypes.SelectMenu:
+    case MessageComponentTypes.SelectMenuChannels:
+    case MessageComponentTypes.SelectMenuRoles:
+    case MessageComponentTypes.SelectMenuUsers:
+    case MessageComponentTypes.SelectMenuUsersAndRoles:
+      return transformSelectMenuComponent(bot, payload)
+  }
+
+  // return {
+  //   type: payload.type,
+  //   custom_id: payload.customId,
+  //   disabled: payload.disabled,
+  //   required: payload.required,
+  //   style: payload.style ? payload.style : undefined,
+  //   label: payload.label,
+  //   emoji: payload.emoji
+  //     ? {
+  //         id: payload.emoji.id?.toString(),
+  //         name: payload.emoji.name,
+  //         animated: payload.emoji.animated,
+  //       }
+  //     : undefined,
+  //   url: payload.url,
+  //   channel_types: payload.channelTypes,
+  //   options: payload.options?.map((option) => ({
+  //     label: option.label,
+  //     value: option.value,
+  //     description: option.description,
+  //     emoji: option.emoji
+  //       ? {
+  //           id: option.emoji.id?.toString(),
+  //           name: option.emoji.name,
+  //           animated: option.emoji.animated,
+  //         }
+  //       : undefined,
+  //     default: option.default,
+  //   })),
+  //   placeholder: payload.placeholder,
+  //   min_values: payload.minValues,
+  //   max_values: payload.maxValues,
+  //   min_length: payload.minLength,
+  //   max_length: payload.maxLength,
+  //   value: payload.value,
+  //   // This casts to a button component, but this really is any component that can go in a actionRow
+  //   components: payload.components?.map((component) => bot.transformers.reverse.component(bot, component)) as DiscordButtonComponent[],
+  //   default_values: payload.defaultValues,
+  //   sku_id: payload.skuId ? bot.transformers.reverse.snowflake(payload.skuId) : undefined,
+  // }
+}
+
+function transformActionRow(bot: Bot, payload: Component): DiscordActionRow {
   return {
-    type: payload.type,
+    type: MessageComponentTypes.ActionRow,
+    // The actionRow.components type is kinda annoying, so we need a cast for this
+    components: (payload.components?.map((component) => bot.transformers.reverse.component(bot, component)) ?? []) as DiscordActionRow['components'],
+  }
+}
+
+function transformButtonComponent(bot: Bot, payload: Component): DiscordButtonComponent {
+  // Since Component is a merge of all components, some casts are necessary
+  return {
+    type: MessageComponentTypes.Button,
+    style: payload.style as ButtonStyles,
     custom_id: payload.customId,
     disabled: payload.disabled,
-    required: payload.required,
-    style: payload.style,
-    label: payload.label,
     emoji: payload.emoji
       ? {
-          id: payload.emoji.id?.toString(),
+          id: payload.emoji.id ? bot.transformers.reverse.snowflake(payload.emoji.id) : undefined,
           name: payload.emoji.name,
           animated: payload.emoji.animated,
         }
       : undefined,
+    label: payload.label,
     url: payload.url,
+    sku_id: payload.skuId ? bot.transformers.reverse.snowflake(payload.skuId) : undefined,
+  }
+}
+
+function transformInputTextComponent(_bot: Bot, payload: Component): DiscordInputTextComponent {
+  // Since Component is a merge of all components, some casts are necessary
+  return {
+    type: MessageComponentTypes.InputText,
+    style: payload.style as TextStyles,
+    custom_id: payload.customId!,
+    label: payload.label!,
+    value: payload.value,
+    max_length: payload.maxLength,
+    min_length: payload.minLength,
+    placeholder: payload.placeholder,
+    required: payload.required,
+  }
+}
+
+function transformSelectMenuComponent(bot: Bot, payload: Component): DiscordSelectMenuComponent {
+  return {
+    type: payload.type as DiscordSelectMenuComponent['type'],
+    custom_id: payload.customId!,
     channel_types: payload.channelTypes,
+    default_values: payload.defaultValues?.map((defaultValue) => ({
+      id: bot.transformers.reverse.snowflake(defaultValue.id),
+      type: defaultValue.type,
+    })),
+    disabled: payload.disabled,
+    max_values: payload.maxValues,
+    min_values: payload.minValues,
     options: payload.options?.map((option) => ({
       label: option.label,
       value: option.value,
       description: option.description,
       emoji: option.emoji
         ? {
-            id: option.emoji.id?.toString(),
+            id: option.emoji.id ? bot.transformers.reverse.snowflake(option.emoji.id) : undefined,
             name: option.emoji.name,
             animated: option.emoji.animated,
           }
@@ -32,11 +136,5 @@ export function transformComponentToDiscordComponent(bot: Bot, payload: Componen
       default: option.default,
     })),
     placeholder: payload.placeholder,
-    min_values: payload.minValues,
-    max_values: payload.maxValues,
-    min_length: payload.minLength,
-    max_length: payload.maxLength,
-    value: payload.value,
-    components: payload.components?.map((component) => bot.transformers.reverse.component(bot, component)),
   }
 }
