+++ packages/bot/src/transformers/reverse/index.ts
@@ -12,7 +12,6 @@ export * from './createApplicationCommand.js'
 export * from './embed.js'
 export * from './emoji.js'
 export * from './gatewayBot.js'
-export * from './interactionResponse.js'
 export * from './member.js'
 export * from './presence.js'
 export * from './team.js'
