+++ packages/utils/src/typeguards.ts
@@ -1,11 +1,14 @@
-import type {
-  DiscordInviteCreate,
-  DiscordInviteMetadata,
-  GetMessagesAfter,
-  GetMessagesAround,
-  GetMessagesBefore,
-  GetMessagesLimit,
-  GetMessagesOptions,
+import {
+  ApplicationCommandTypes,
+  type CreateApplicationCommand,
+  type CreateContextApplicationCommand,
+  type DiscordInviteCreate,
+  type DiscordInviteMetadata,
+  type GetMessagesAfter,
+  type GetMessagesAround,
+  type GetMessagesBefore,
+  type GetMessagesLimit,
+  type GetMessagesOptions,
 } from '@discordeno/types'
 import { hasProperty } from './utils.js'
 
@@ -27,4 +30,8 @@ export function isGetMessagesLimit(options: GetMessagesOptions): options is GetM
 
 export function isInviteWithMetadata(options: DiscordInviteCreate | DiscordInviteMetadata): options is DiscordInviteMetadata {
   return !hasProperty(options, 'channel_id')
+}
+
+export function isContextApplicationCommand(command: CreateApplicationCommand): command is CreateContextApplicationCommand {
+  return command.type === ApplicationCommandTypes.Message || command.type === ApplicationCommandTypes.User
 }
